# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Torneo Solidario",
        "description": "API del torneo solidario",
        "contact": {
            "email": "minillanillo@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "registrar a un usuario en la web",
                "description": "Registrar a un usuario en la web para devolver su rol y su token con abilities",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Datos del usuario a loguearphp",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "minillanillo@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario logueado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario logueado correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Usuario"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales incorrectas. Acceso no autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Usuario": {
                "required": [
                    "nombre",
                    "email",
                    "rol"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "David"
                    },
                    "email": {
                        "type": "string",
                        "example": "Merchandising de la Liga"
                    },
                    "rol": {
                        "type": "string",
                        "example": "[entrenador | director | periodista | administrador]"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to authenticate",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}